Move legacy htmlParser source into AngularDirectiveParser
All of this logic needs to be rewriteen to work with the newer plugin
system and better AST walking.
