+++ packages/hekla-core/src/parsers/AngularDirectiveParser/index.js
@@ -1,11 +1,10 @@
 'use strict';
 
 const path = require('path');
+const htmlparser = require('htmlparser2');
 
 const astUtils = require('../../utils/ast-utils');
-const fsUtils = require('../../utils/fs-utils');
 const ngUtils = require('../../utils/ng-utils');
-const htmlParser = require('../../analyzer/html-parser');
 const BaseParser = require('../BaseParser');
 const ParserResult = require('../../utils/parser-result');
 
@@ -130,7 +129,7 @@ function getDependencies(templateInfo) {
 
   // console.log('getting dependencies:', templateInfo);
   return Promise.resolve()
-    .then(() => htmlParser.getDependencies(templateInfo.contents, templateInfo.path))
+    .then(() => getDependenciesLegacy(templateInfo.contents, templateInfo.path))
     .catch(err => {
       console.log('dependency error!', err);
       if (err.code === 'ENOENT') {
@@ -142,7 +141,62 @@ function getDependencies(templateInfo) {
     });
 }
 
-function debugNode(node) {
-  // TODO: remove this function
-  console.log(JSON.stringify(node, null, 2));
+function getDependenciesLegacy(fileContents, filePath) {
+  const dependencies = [];
+
+  var parser = new htmlparser.Parser({
+
+    onopentag: (name, attrs) => {
+      if (isSpecialTag(name)) {
+        dependencies.push(name);
+      }
+
+      // TODO: check special attributes
+      // for (let key in attrs) {
+      // }
+    }
+
+  }, {decodeEntities: true});
+  parser.write(fileContents);
+  parser.end();
+
+  return cleanDependencyList(dependencies);
+}
+
+function isSpecialTag(tagName) {
+  return (tagName.indexOf('-') !== -1);
+}
+
+function getDmComponentName(attrValue, filePath) {
+  let name = attrValue;
+
+  if (attrValue.indexOf('./') === 0) {
+    const pathPieces = filePath.split('/');
+    const folderName = pathPieces[pathPieces.length - 2];
+    name = name.replace('./', folderName + '/');
+  }
+
+  return name;
+}
+
+function cleanDependencyList(dependencies) {
+  const uniqueDependencies = uniq(dependencies);
+  return uniqueDependencies
+    .sort();
+}
+
+// Courtesy of http://stackoverflow.com/a/9229821/2418448
+function uniq(a) {
+  var seen = {};
+  var out = [];
+  var len = a.length;
+  var j = 0;
+  for(var i = 0; i < len; i++) {
+    var item = a[i];
+    if(seen[item] !== 1) {
+      seen[item] = 1;
+      out[j++] = item;
+    }
+  }
+  return out;
 }
